reuse shared topk1 chain buffers; drop per-bs cache
